refactor(ci): drop the typecheck pass the measurement showed was redundant
#419 asked what each typecheck pass catches that the others do not. Measured by
injecting a deliberate type error into each area and recording which passes went
red: eight of the ten are uniquely responsible for something, and one is not.

`contributing:typecheck` compiled the twelve fixtures under
`test/some-code-from-docs/contributing/` alone. Since #428 widened
`test/tsconfig.json` to the whole test tree those files are in
`test:typecheck`'s program too, under stricter settings — that config extends the
package one (`strict: true`) while the fixture config set
`strict: false, noImplicitAny: false`. The redundancy did not exist when #419 was
written; the fix for #428 created it.

Verified by injection rather than by reading configs: a type error in a fixture
still fails `test:typecheck` with the pass gone, and the fresh-clone guard still
prints its `dev:prepare` hint.

That guard is the one thing the removal could have lost, so `contributing-typecheck.mjs`
became `test-typecheck.mjs` — same `requireSdkTypes` call, pointed at
`test/tsconfig.json` — rather than being deleted. The fixtures' own `tsconfig.json`
goes with the pass; their `.spec.ts` files still run in the `jsSdk:unit` project
and are untouched.

Nine passes, nine hand-written tsconfigs. `testing.md` now carries the measured
table of what each pass uniquely covers, so the next person does not re-measure,
along with the two results worth keeping: the playground passes are not redundant
(an error in either playground is caught by its own pass and nothing else), and
the code fences in `.github/contributing/*.md` are compiled by nothing (#435).

Refs #419
