+++ package.json
@@ -25,7 +25,7 @@
     "package-jssdk:build": "pnpm --filter ./packages/jssdk build",
     "package-jssdk:lint": "pnpm --filter ./packages/jssdk lint",
     "package-jssdk:typecheck": "pnpm --filter ./packages/jssdk typecheck",
-    "test:typecheck": "pnpm run skills:install && tsc --noEmit -p test/tsconfig.json",
+    "test:typecheck": "pnpm run skills:install && node scripts/test-typecheck.mjs",
     "package-jssdk-nuxt:prepare": "pnpm --filter ./packages/jssdk-nuxt dev:prepare",
     "package-jssdk-nuxt:build": "pnpm --filter ./packages/jssdk-nuxt build",
     "package-jssdk-nuxt:lint": "pnpm --filter ./packages/jssdk-nuxt lint",
@@ -55,7 +55,6 @@
     "package-jssdk:test": "vitest --project jsSdk:integration",
     "package-jssdk:test:run": "vitest run --project jsSdk:integration --project jsSdk:contributing-snippets",
     "package-jssdk:test-contributing-snippets": "vitest run --project jsSdk:contributing-snippets",
-    "contributing:typecheck": "node scripts/contributing-typecheck.mjs",
     "package-jssdk:test-integration-js-docs": "vitest -t \"js-docs\" --project jsSdk:integration",
     "package-jssdk:test-integration-core": "vitest -t \"core\" --project jsSdk:integration",
     "package-jssdk:test:run-underLoad-v3-call": "vitest run -t \"tasksTaskGet @apiV3\" --project jsSdk:underLoad",
@@ -77,7 +76,7 @@
     "docs:lint-links": "node scripts/docs-link-check.mjs",
     "docs:lint-api-index": "node scripts/check-api-reference-index.mjs",
     "docs:lint:test": "node --test \"scripts/__tests__/**/*.test.mjs\"",
-    "typecheck": "pnpm run package-jssdk:typecheck && pnpm run test:typecheck && pnpm run package-jssdk-nuxt:typecheck && pnpm run docs:typecheck && pnpm run playground-nuxt:typecheck && pnpm run playground-cli:typecheck && pnpm run docs:typecheck-blocks && pnpm run skills:typecheck && pnpm run skills:typecheck-blocks && pnpm run contributing:typecheck",
+    "typecheck": "pnpm run package-jssdk:typecheck && pnpm run test:typecheck && pnpm run package-jssdk-nuxt:typecheck && pnpm run docs:typecheck && pnpm run playground-nuxt:typecheck && pnpm run playground-cli:typecheck && pnpm run docs:typecheck-blocks && pnpm run skills:typecheck && pnpm run skills:typecheck-blocks",
     "release:bump": "node scripts/bump-version.mjs"
   },
   "devDependencies": {
