--- test/some-code-from-docs/contributing/tsconfig.json
@@ -1,21 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "compilerOptions": {
-    "target": "ESNext",
-    "module": "ESNext",
-    "moduleResolution": "bundler",
-    "lib": ["ESNext", "DOM"],
-
-    "strict": false,
-    "noImplicitAny": false,
-    "skipLibCheck": true,
-    "noEmit": true,
-
-    "moduleDetection": "force",
-
-    "allowJs": false,
-    "isolatedModules": true
-  },
-  "include": ["./**/*.ts"],
-  "exclude": ["./**/*.spec.ts"]
-}
